requirements.txt
tensorflow-gpu==1.11.0
stanford-corenlp
tqdm
h5py
ujson